+++ rustfmt.toml
@@ -1,0 +1,3 @@
+reorder_imports = true
+group_imports = 'StdExternalCrate'
+imports_granularity = "Crate"
